add draft KPI for link channel centre id checking (#143)
* add draft KPI for link channel centre id checking

* fix tests

* Add sanity checks to centre-id KPIs #143

Expecting a list to contain the desired number of elements is somewhat naive; doing it twice is unforgivable :-)
One of the test examples already ended with an index out of bounds error, added another that illustrates the other list issue.

* Add forgotten file sanity checks to centre-id KPIs #143

Did too many experiments and finally forgot to commit the actual changed file.

* Add example file for the sanity checks of centre-id #143

* Fix the new unit test #143

What about saving the file before commit?

* Please the flake8 #143

---------

Co-authored-by: Jan Osusky <Jan.Osusky@iblsoft.com>